railway-yolo-detection/requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Dataset download
roboflow>=1.1.0
requests>=2.28.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
